requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
textblob==0.17.1
nltk==3.8.1
requests==2.31.0

